The file TeamListHook has been renamed by staff and relocated in the correct path src/component/team #931
